+++ tests/modeling/test_modeling.py
@@ -19,6 +19,7 @@ from pharmpy.modeling import (
     create_joint_distribution,
     fix_parameters_to,
     has_first_order_elimination,
+    has_linear_odes,
     has_michaelis_menten_elimination,
     has_mixed_mm_fo_elimination,
     has_odes,
@@ -3536,3 +3537,13 @@ def test_has_odes(load_example_model_for_test, datadir, load_model_for_test):
     path = datadir / 'minimal.mod'
     model = load_model_for_test(path)
     assert not has_odes(model)
+
+
+def test_has_linear_odes(load_example_model_for_test, datadir, load_model_for_test):
+    model = load_example_model_for_test('pheno')
+    assert has_linear_odes(model)
+    model = set_michaelis_menten_elimination(model)
+    assert not has_linear_odes(model)
+    path = datadir / 'minimal.mod'
+    model = load_model_for_test(path)
+    assert not has_linear_odes(model)